Rename module to `commands.py`
NHUB-546
