+++ newsroom/commands/scheduled_notifications.py
@@ -1,11 +1,10 @@
 import click
-from quart.cli import with_appcontext
 
-from newsroom.notifications.send_scheduled_notifications import SendScheduledNotificationEmails
+from newsroom.notifications.commands import SendScheduledNotificationEmails
 from .cli import newsroom_cli
 
 
-@newsroom_cli.command("send_scheduled_notifications")
+@newsroom_cli.register_async_command("send_scheduled_notifications", with_appcontext=True)
 @click.option(
     "-i",
     "--ignore-schedule",
@@ -14,8 +13,7 @@ from .cli import newsroom_cli
     required=False,
     help="Runs a schedule if one has not been run for that user's schedule",
 )
-@with_appcontext
-def send_scheduled_notifications(force=False):
+async def send_scheduled_notifications(force=False):
     """
     Send scheduled notifications
 
@@ -24,4 +22,4 @@ def send_scheduled_notifications(force=False):
 
         $ python manage.py send_scheduled_notifications
     """
-    SendScheduledNotificationEmails().run(force)
+    await SendScheduledNotificationEmails().run(force)
